delete splash navigation and replace by main_graph.xml. add HomeFragment.kt and check navigation functionality
