+++ app/src/main/AndroidManifest.xml
@@ -17,20 +17,12 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".ui.activities.InitialHostActivity"
+            android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-
-        <activity
-            android:name=".MainActivity"
-            android:exported="false">
         </activity>
 
     </application>
