--- app/src/main/java/com/Alkemy/alkemybankbase/ui/activities/InitialHostActivity.kt
@@ -1,12 +1,0 @@
-package com.Alkemy.alkemybankbase.ui.activities
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.Alkemy.alkemybankbase.R
-
-class InitialHostActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_initial_host)
-    }
-}
