+++ app/src/main/res/navigation/main_graph.xml
@@ -2,9 +2,14 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/login_register_graph"
+    android:id="@+id/main_graph"
     app:startDestination="@id/splashFragment">
 
+    <fragment
+        android:id="@+id/homeFragment"
+        android:name="com.Alkemy.alkemybankbase.ui.fragments.HomeFragment"
+        android:label="fragment_home"
+        tools:layout="@layout/fragment_home" />
     <fragment
         android:id="@+id/splashFragment"
         android:name="com.Alkemy.alkemybankbase.ui.fragments.SplashFragment"
@@ -18,5 +23,9 @@
         android:id="@+id/loginFragment"
         android:name="com.Alkemy.alkemybankbase.ui.fragments.LoginFragment"
         android:label="fragment_login"
-        tools:layout="@layout/fragment_login" />
+        tools:layout="@layout/fragment_login" >
+        <action
+            android:id="@+id/action_loginFragment_to_homeFragment"
+            app:destination="@id/homeFragment" />
+    </fragment>
 </navigation>
